Move CacheStorageAccessor and StaticAssetService to Demo.Shared
