+++ src/Core/Infrastructure/LibraryConfiguration.cs
@@ -7,10 +7,6 @@
 /// </summary>
 public class LibraryConfiguration
 {
-    public BlazorHostingModel HostingModel { get; set; } = BlazorHostingModel.NotSpecified;
-
-    public StaticAssetServiceConfiguration StaticAssetServiceConfiguration { get; set; } = new();
-
     /// <summary>
     /// Gets or sets a value indicating whether the library should use the TooltipServiceProvider.
     /// If set to true, add the FluentTooltipProvider component at end of the MainLayout.razor page.
